% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAnswerSchema.R
\name{GetAnswerSchema}
\alias{GetAnswerSchema}
\title{Get the answer schema of a form.}
\usage{
GetAnswerSchema(token, idUser, idForm)
}
\arguments{
\item{token}{A string access token.}

\item{idUser}{Numeric Id of the user.}

\item{idForm}{Numeric Id of the required form.}
}
\value{
A possible nested data frame.
}
\description{
Get the schema of the answers of a especific form in the shape a nested data
frame, that contain all the needed information to request the answers of the
form.
}
\examples{
GetAnswerSchema('eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9', 1, 3345)
}
