% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFormSchema.R
\name{GetFormSchema}
\alias{GetFormSchema}
\title{Get the form schema of a form.}
\usage{
GetFormSchema(token, idAccount, idForm)
}
\arguments{
\item{token}{A string access token.}

\item{idAccount}{Numeric Id of the account.}

\item{idForm}{Numeric Id of the required form.}
}
\value{
A possible nested data frame.
}
\description{
Get the schema of the questions of a especific form in the shape a nested data
frame, that contain all the needed information to request the answers of the
form.
}
\examples{
GetFormSchema('eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9', 1, 3345)
}
