% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAnswers.R
\name{GetAnswers}
\alias{GetAnswers}
\title{Get all the answers of a form.}
\usage{
GetAnswers(token, idUser, idForm)
}
\arguments{
\item{token}{A string access token.}

\item{idUser}{Numeric Id of the user.}

\item{idForm}{Numeric Id of the required form.}
}
\value{
A data frame.
}
\description{
Get all the currents answers of a especific form. This function make a call
to GetAnswerSchema().
}
\examples{
GetAnswers('eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9', 1, 3345)
}
