% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAnswers.R
\name{GetAnswers}
\alias{GetAnswers}
\title{Get all the answers of a form.}
\usage{
GetAnswers(token, idAccount, idForm, formSource = NULL,
  createdBefore = NULL, createdAfter = NULL)
}
\arguments{
\item{token}{A string access token.}

\item{idAccount}{Numeric Id of the account.}

\item{idForm}{Numeric Id of the required form.}

\item{formSource}{Optional filter. Is the origin of the source of the answer
Can use 'web_public', 'web_private' and 'mobile'.}

\item{createdBefore}{Optional filter. This parameter filter the answers
that was answered before this date. Is acceptable the ISO8601 format
("YYYY-MM-DD"). Also is possible specify another format, sending together
in a vector in the R especification, for example, "%d-%m-%Y" to "25-10-1995".}

\item{createdAfter}{Optional filter. This parameter filter the answers
that was answered after this date. Is acceptable the ISO8601 format
("YYYY-MM-DD"). Also is possible specify another format, sending together
in a vector in the R especification, for example, "%d-%m-%Y" to "25-10-1995".}
}
\value{
A data frame.
}
\description{
Get all the currents answers of a especific form. This function make a call
to GetFormSchema().
}
\examples{
GetAnswers('eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9', 1, 3345)
}
