% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAnswers.R
\name{GetAnswers}
\alias{GetAnswers}
\title{Get all the answers of a form.}
\usage{
GetAnswers(token, idAccount, idForm, repetedColunsNames = FALSE,
  formSource = NULL, createdBefore = NULL, createdAfter = NULL)
}
\arguments{
\item{token}{A string access token.}

\item{idAccount}{Numeric Id of the account.}

\item{idForm}{Numeric Id of the required form.}

\item{repetedColunsNames}{Boolean flag, indicates if the repeted columns
names will stay or if gonna be rename with a suffix.}

\item{formSource}{Optional filter. Is the origin of the source of the answer
Can use 'web_public', 'web_private' and 'mobile'.}

\item{createdBefore}{Optional filter. This parameter filter the answers
that was answered before this date. Is acceptable the ISO8601 format
("YYYY-MM-DD"). Also is possible specify another format, sending together
in a vector in the R especification, for example, "%d-%m-%Y" to "25-10-1995".}

\item{createdAfter}{Optional filter. This parameter filter the answers
that was answered after this date. Is acceptable the ISO8601 format
("YYYY-MM-DD"). Also is possible specify another format, sending together
in a vector in the R especification, for example, "%d-%m-%Y" to "25-10-1995".}
}
\value{
A list, with one or more data frames.
}
\description{
Get all the currents answers of a specific form. This function make a call
to GetFormSchema().
}
\examples{
GetAnswers('eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9', 1, 3345)
GetAnswers(token = token,idAccount = 1,
             idForm = idForm,
             formSource = NULL,
             createdAfter = "2012-12-20",
             createdBefore = c("20-12-2018","\%d-\%m-\%Y")
             )
GetAnswers(token = token,idAccount = 1,
             idForm = idForm,
             formSource = web_public,
             createdAfter = c("20-12-2012","\%d-\%m-\%Y"),
             createdBefore = c("20-12-2018","\%d-\%m-\%Y")
             )
GetAnswers(token = token,idAccount = 1,
             idForm = idForm,
             formSource = web_private,
             createdAfter = "2012-12-20",
             createdBefore = "2018-12-20",
             )
}
