% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAnswers.R
\name{GetAnswers}
\alias{GetAnswers}
\title{Get all the answers of a form.}
\usage{
GetAnswers(token, idForm, nameForm, repetedColunsNames = FALSE,
  formSource = NULL, createdAfter = NULL, createdBefore = NULL)
}
\arguments{
\item{token}{String access token.}

\item{idForm}{Numeric Id of the required form.}

\item{nameForm}{String name of the required form. Just is used when an idForm
is not supplied. When this parameter is used, are spent extra one access
quota.}

\item{repetedColunsNames}{Boolean flag, indicates if the repeated columns
names will stay or if gonna be renamed with a suffix.}

\item{formSource}{Optional filter. Is the origin of the source of the answer
can use 'web_public', 'web_private' and 'mobile'.}

\item{createdAfter}{Optional filter. This parameter filters the answers that
were answered after this date. Is acceptable in the ISO8601 format
("YYYY-MM-DD"). Also is possible to specify another format, sending together
in a vector in the R especification, for example, "\%d-\%m-\%Y" to
"25-10-1995".}

\item{createdBefore}{Optional filter. This parameter filters the answers
that were answered before this date. Is acceptable in the ISO8601 format
("YYYY-MM-DD"). Also is possible to specify another format, sending together
in a vector in the R especification, for example, "\%d-\%m-\%Y" to
"25-10-1995".}
}
\value{
A list, with one or more data frames.
}
\description{
Get all the currents answers of a specific form. This function makes a call
to GetFormStructure and spent 2 quotas.
}
\examples{
\dontrun{
GetAnswers('eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9', 3345)
GetAnswers(token = token,
             idForm = idForm,
             formSource = NULL,
             createdAfter = "2012-12-20",
             createdBefore = c("20-12-2018","\%d-\%m-\%Y")
             )
GetAnswers(token = token,
             idForm = idForm,
             formSource = web_public,
             createdAfter = c("20-12-2012","\%d-\%m-\%Y"),
             createdBefore = c("20-12-2018","\%d-\%m-\%Y")
             )
GetAnswers(token = token,
             idForm = idForm,
             formSource = web_private,
             createdAfter = "2012-12-20",
             createdBefore = "2018-12-20",
             )
}
}
