% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAnswers.R
\name{GetAnswers}
\alias{GetAnswers}
\title{Get all the answers of a form.}
\usage{
GetAnswers(token, idAccount, idForm)
}
\arguments{
\item{token}{A string access token.}

\item{idAccount}{Numeric Id of the account.}

\item{idForm}{Numeric Id of the required form.}
}
\value{
A data frame.
}
\description{
Get all the currents answers of a especific form. This function make a call
to GetFormSchema().
}
\examples{
GetAnswers('eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9', 1, 3345)
}
