% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAnswers.R
\name{GetAnswers}
\alias{GetAnswers}
\title{Get all the answers of a form.}
\usage{
GetAnswers(token, idForm, nameForm = NULL, singleDataFrame = FALSE,
  source = NULL, createdAfter = NULL, createdBefore = NULL)
}
\arguments{
\item{token}{String access token.}

\item{idForm}{Numeric Id of the required form.}

\item{nameForm}{String name of the required form. Just is used when an idForm
is not supplied. When this parameter is used, are spent extra one access
quota.}

\item{singleDataFrame}{Boolean flag. Indicates the preference to create a
single data frame with all the answers. In this case, is possible to have
repeated values, according to the multiplicity of relationships.}

\item{source}{Optional filter. Is the the source of the answer and can use
'web_public', 'web_private' or 'mobile'.}

\item{createdAfter}{Optional filter. This parameter filters the answers that
were answered after this date. Is acceptable in the ISO8601 format
("YYYY-MM-DD"). Also is possible to specify another format, sending together
in a vector in the R especification, for example, "\%d-\%m-\%Y" to
"25-10-1995".}

\item{createdBefore}{Optional filter. This parameter filters the answers
that were answered before this date. Is acceptable in the ISO8601 format
("YYYY-MM-DD"). Also is possible to specify another format, sending together
in a vector in the R especification, for example, "\%d-\%m-\%Y" to
"25-10-1995".}
}
\value{
A list, with one or more data frames.
}
\description{
Get all the currents answers of a specific form. This function makes a call
to GetFormStructure and spent 2 quotas.
}
\details{
To get more details about the fields provided by the result, please visit the
\href{https://coletum.docs.apiary.io/#reference/0/preenchimentos/listar-preenchimentos}{API documentation}.
}
\examples{
\dontrun{
GetAnswers('eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9', 3345)
GetAnswers('eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9', ,'form 123', TRUE)
GetAnswers(token = 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9',
             nameForm = 'form 123')
GetAnswers(token = token,
             idForm = idForm,
             source = NULL,
             createdAfter = "2012-12-20",
             createdBefore = c("20-12-2018","\%d-\%m-\%Y")
             )
GetAnswers(token = token,
             idForm = idForm,
             source = web_public,
             createdAfter = c("20-12-2012","\%d-\%m-\%Y"),
             createdBefore = c("20-12-2018","\%d-\%m-\%Y")
             )
GetAnswers(token = token,
             idForm = idForm,
             singleDataFrame = TRUE,
             source = web_private,
             createdAfter = "2012-12-20",
             createdBefore = "2018-12-20",
             )
}
}
