% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetForms.R
\name{GetForms}
\alias{GetForms}
\title{Get infos of all forms.}
\usage{
GetForms(token, idUser)
}
\arguments{
\item{token}{A string access token.}

\item{idUser}{Numeric Id of the user.}
}
\value{
A data frame.
}
\description{
Get the principals infos of all forms.
}
\examples{
GetForms('eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9', 5)
}
