% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetForms.R
\name{GetForms}
\alias{GetForms}
\title{Get info of all forms.}
\usage{
GetForms(token, status = NULL, public_answers = NULL,
  answer_tracking = NULL)
}
\arguments{
\item{token}{String access token.}

\item{status}{Optional filter. That is the state of the form: accept
'enabled' or 'disabled'.}

\item{public_answers}{Optional filter. If the form accepts answers
anonymously, are possible to use 'true' or 'false'.}

\item{answer_tracking}{Optional filter. If the form is saving the local of
fill, is possible use 'true' or 'false'.}
}
\value{
A data frame.
}
\description{
Get the principals info of all forms.
}
\examples{
\dontrun{
GetForms('eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9')
GetForms('eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9','enabled','true','true')
GetForms(token = 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9',
         status = 'enabled',
         public_answers = 'false',
         answer_tracking = 'true'
         )
}
}
