% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RColetum.r
\docType{package}
\name{RColetum-package}
\alias{RColetum}
\alias{RColetum-package}
\title{RColetum}
\description{
An R package to get data from \href{https://coletum.com}{Coletum}.
}
\section{\pkg{RColetum} functions}{

The currents functions that are provide by \pkg{RColetum} are:
\itemize{

\item \code{\link{Auth}}: is used to authenticate the token and return the
user account infos.
\item \code{\link{Logoff}}: destroy the credentials of the token.
\item \code{\link{GetForms}}: get the information about all forms or the
especificate one.
\item \code{\link{GetAnswerSchema}}: get the answer schema of a form.
\item \code{\link{GetAnswers}}: get the all the answers of a form or
personalize with filters and take just what you need.
\item \code{\link{GetAccountInfos}}: get important account infos to use the
other functions.

}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/geo-sapiens/RColetum}
}

}
