% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RColetum.r
\docType{package}
\name{RColetum-package}
\alias{RColetum}
\alias{RColetum-package}
\title{RColetum}
\description{
An R package to get data from \href{https://coletum.com}{Coletum}.
}
\section{\pkg{RColetum} functions}{

The currents functions that are provide by \pkg{RColetum} are:
\itemize{

\item \code{\link{GetForms}}: get the information about all forms.
\item \code{\link{GetFormSchema}}: get the form schema of a form.
\item \code{\link{GetAnswers}}: get the all the answers of a form or
personalize with filters and take just what you need.

}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/geo-sapiens/RColetum}
  \item Report bugs at \url{https://github.com/geo-sapiens/RColetum/issues}
}

}
\author{
\strong{Maintainer}: André Smaniotto \email{smaniotto@geosapiens.com}

Other contributors:
\itemize{
  \item GeoSapiens [copyright holder, funder]
}

}
